requirements.txt
discord.py
flask
python-dotenv
werkzeug